+++ samples/import-export/import-environment.rb
@@ -442,7 +442,6 @@ if options.importCore
       end
 
       # Import Submissions
-      # Not using Parallel here because unlicensed server will fail at 25 counts
       submissions_count = 0
       Dir["#{dirname}/data/*"].each do |form|
         # Parse form slug from directory path